requirements.txt

pandas
scikit-learn
azureml-sdk
